LEP-222 redact existing dates of birth in logs
