+++ lib/tasks/redact_data.rake
@@ -1,6 +1,6 @@
 namespace :db do
   desc "batch job to redact client_id's in the request_log requests for payments"
-  task redact_client_ids: :environment do
-    RedactClientIdService.call
+  task redact_data: :environment do
+    RedactService.call
   end
 end
